Fix Order Post
Fix Order Post, change DTOs,  Mappers and responses.

+++ backend/src/main/java/com/nocountry/backend/repository/IUserReviewRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface IUserReviewRepository extends JpaRepository<UserReview, Integer> {
+public interface IUserReviewRepository extends JpaRepository<UserReview, Long> {
 }
